asciidoc-sections: strip separator prefix from section anchor link
This should help fix one of the more annoying issues with the jenkins
pipeline step docs where the step links never work. If the title
contains invalid section id chars at the begging of the string the result
would be a leading separator char.

For example if the title is "<code>foo</code>: bar" the resulting anchor
should be "foo-bar" not "-foo-bar".